requirements.txt
streamlit
supabase
google-generativeai
langchain
langchain-google-genai
langchain-community
pandas
python-dotenv
sqlalchemy
psycopg2-binary
